+++ truffle/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/generator/FlatNodeGenFactory.java
@@ -61,6 +61,7 @@ import javax.lang.model.type.TypeMirror;
 
 import com.oracle.truffle.api.Assumption;
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
+import com.oracle.truffle.api.dsl.Reflection.ReflectedSpecialization.State;
 import com.oracle.truffle.api.dsl.UnsupportedSpecializationException;
 import com.oracle.truffle.api.dsl.internal.DSLOptions;
 import com.oracle.truffle.api.nodes.ExplodeLoop;
@@ -87,6 +88,7 @@ import com.oracle.truffle.dsl.processor.java.model.CodeTree;
 import com.oracle.truffle.dsl.processor.java.model.CodeTreeBuilder;
 import com.oracle.truffle.dsl.processor.java.model.CodeTypeElement;
 import com.oracle.truffle.dsl.processor.java.model.CodeTypeMirror.ArrayCodeTypeMirror;
+import com.oracle.truffle.dsl.processor.java.model.CodeTypeMirror.DeclaredCodeTypeMirror;
 import com.oracle.truffle.dsl.processor.java.model.CodeVariableElement;
 import com.oracle.truffle.dsl.processor.model.AssumptionExpression;
 import com.oracle.truffle.dsl.processor.model.CacheExpression;
@@ -364,7 +366,102 @@ public class FlatNodeGenFactory {
 
         clazz.getEnclosedElements().addAll(removeThisMethods.values());
 
+        if (node.isReflectable()) {
+            generateReflectionInfo(clazz);
+        }
+
         return clazz;
+    }
+
+    private void generateReflectionInfo(CodeTypeElement clazz) {
+        clazz.getImplements().add(context.getType(ReflectionAccesor.class));
+        CodeExecutableElement reflection = new CodeExecutableElement(modifiers(PUBLIC), context.getType(Object.class), "getReflectionData");
+
+        CodeTreeBuilder builder = reflection.createBuilder();
+
+        List<SpecializationData> filteredSpecializations = new ArrayList<>();
+        for (SpecializationData s : node.getSpecializations()) {
+            if (s.getMethod() == null) {
+                continue;
+            }
+            filteredSpecializations.add(s);
+        }
+
+        ArrayCodeTypeMirror objectArray = new ArrayCodeTypeMirror(context.getType(Object.class));
+        builder.declaration(objectArray, "data", builder.create().startNewArray(objectArray, CodeTreeBuilder.singleString(String.valueOf(filteredSpecializations.size()))).end().build());
+        builder.declaration(objectArray, "s", (CodeTree) null);
+
+        FrameState frameState = FrameState.load(this);
+        builder.tree(state.createLoad(frameState, null));
+        if (requiresExclude()) {
+            builder.tree(exclude.createLoad(frameState, null));
+        }
+
+        int index = 0;
+        for (SpecializationData specialization : filteredSpecializations) {
+            builder.startStatement().string("s = ").startNewArray(objectArray, CodeTreeBuilder.singleString("3")).end().end();
+            builder.startStatement().string("s[0] = ").doubleQuote(specialization.getMethodName()).end();
+            builder.startIf().tree(state.createContains(frameState, new Object[]{specialization})).end().startBlock();
+            builder.startStatement().string("s[1] = ").staticReference(context.getType(State.class), "ACTIVE").end();
+
+            TypeMirror listType = new DeclaredCodeTypeMirror((TypeElement) context.getDeclaredType(ArrayList.class).asElement(), Arrays.asList(context.getType(Object.class)));
+
+            if (!specialization.getCaches().isEmpty()) {
+                builder.declaration(listType, "cached", "new ArrayList<>()");
+
+                boolean useSpecializationClass = useSpecializationClass(specialization);
+
+                String name = createSpecializationLocalName(specialization);
+
+                if (useSpecializationClass) {
+                    builder.declaration(createSpecializationTypeName(specialization), name, CodeTreeBuilder.singleString(createSpecializationFieldName(specialization)));
+
+                    if (specialization.getMaximumNumberOfInstances() > 1) {
+                        builder.startWhile();
+                    } else {
+                        builder.startIf();
+                    }
+                    builder.string(name, " != null");
+                    builder.end();
+                    builder.startBlock();
+                }
+
+                builder.startStatement().startCall("cached", "add");
+                builder.startStaticCall(context.getType(Arrays.class), "asList");
+                for (CacheExpression cache : specialization.getCaches()) {
+                    builder.startGroup();
+                    builder.tree(createCacheReference(specialization, cache.getParameter()));
+                    builder.end();
+                }
+                builder.end();
+                builder.end().end();
+
+                if (useSpecializationClass) {
+                    if (specialization.getMaximumNumberOfInstances() > 1) {
+                        builder.startStatement().string(name, " = ", name, ".next_").end();
+                    }
+
+                    builder.end(); // cache while or if
+                }
+
+                builder.statement("s[2] = cached");
+            }
+            builder.end();
+            if (mayBeExcluded(specialization)) {
+                builder.startElseIf().tree(exclude.createContains(frameState, new Object[]{specialization})).end().startBlock();
+                builder.startStatement().string("s[1] = ").staticReference(context.getType(State.class), "EXCLUDED").end();
+                builder.end();
+            }
+            builder.startElseBlock();
+            builder.startStatement().string("s[1] = ").staticReference(context.getType(State.class), "INACTIVE").end();
+            builder.end();
+            builder.startStatement().string("data[", String.valueOf(index), "] = s").end();
+            index++;
+        }
+
+        builder.startReturn().string("data").end();
+
+        clazz.add(reflection);
     }
 
     private void createFields(CodeTypeElement clazz) {